requirements.txt
PySide6>=6.6
python-mpv>=1.0
requests>=2.31
python-dotenv>=1.0
